refactor: file validation
