+++ src/constants/files.js
@@ -1,0 +1,9 @@
+export const MAX_FILE_SIZE = 5 * 1024 * 1024;
+
+export const FileType = {
+  IMAGE: 'image/',
+  TEXT: 'text/plain',
+  DOC: 'application/msword',
+  DOCX: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+  PDF: 'application/pdf',
+};
