+++ src/helpers/deleteFile.js
@@ -1,3 +1,5 @@
+import { toast } from 'react-toastify';
+
 import axios from 'axios';
 
 import { getPresignedDeleteURL } from '@/api/getPresignedDeleteUrl';
@@ -11,8 +13,12 @@ export const deleteFile = async (fileName, category) => {
     await axios.delete(presignedUrl);
 
     console.log('File deleted successfully');
+
+    // toast.success('File deleted successfully');
   } catch (error) {
     console.error('Error deleting file:', error);
+
+    toast.error('Error deleting file');
 
     throw error;
   }
